# Improve TrinomialTree performance with entropy-based pruning under small time steps