# Entropy-Based Pruning in Trinomial Tree Lattices for Efficient Hull-White Pricing

**Author**: Prem Dev\
**Date**: July 2, 2025

---

## Abstract

In affiliation to issue #364

This document proposes an enhancement to QuantLib's `TrinomialTree` class to improve computational efficiency when the time discretization step is very small—a common scenario in pricing interest rate derivatives like Bermudan swaptions under the Hull-White model. The patch introduces entropy-based pruning to suppress unnecessary branching in nearly deterministic transitions. Benchmarks show an 8.6% increase in throughput without any test regressions.

---

## 1. Context and Motivation

The Hull-White one-factor model assumes the short rate \$r\_t\$ evolves as an Ornstein-Uhlenbeck process:

$$
dr_t = (\theta(t) - a r_t) dt + \sigma dW_t
$$

Where:

- \$a\$: mean-reversion rate
- \$\sigma\$: volatility
- \$\theta(t)\$: ensures fit to initial term structure

For small \$\Delta t\$:

$$
\mathbb{E}[r_{t+\Delta t}] = r_t e^{-a \Delta t} + \frac{\theta(t)}{a}(1 - e^{-a \Delta t})
$$

$$
\text{Var}[r_{t+\Delta t}] = \frac{\sigma^2}{2a} (1 - e^{-2a \Delta t})
$$

As \$\Delta t \to 0\$, variance collapses. However, QuantLib’s tree expands nodes based on high branching even when the process becomes deterministic.

---

## 2. Observed Issue in QuantLib

The state spacing in `TrinomialTree` is:

$$
\Delta x_i = \sqrt{3 \cdot \text{Var}_i}
$$

Node index is:

$$
\texttt{temp} = \left\lfloor \frac{\mathbb{E}[r_{t+\Delta t}] - x_0}{\Delta x_{i+1}} + 0.5 \right\rfloor
$$

Problems when \$\Delta t\$ is small:

- \$\Delta x \to 0\$
- `temp` becomes large
- Creates many unused nodes
- Correct but inefficient

---

## 3. Proposed Fix: Entropy-Based Pruning

We compute Shannon entropy:

$$
H = -\sum_{i=1}^{3} p_i \log(p_i)
$$

### Pruning Criteria

```cpp
if (v2 < 1e-12) {
    branching.add(temp, 0.0, 1.0, 0.0);
    continue;
}

Real H = entropy(p1, p2, p3);
if (H < 1e-3 && std::max(p1, p3) < 1e-12) {
    p1 = 0.0; p2 = 1.0; p3 = 0.0;
}
branching.add(temp, p1, p2, p3);
```

### Interpretation

- If transition is deterministic, enforce a single middle-node jump
- Eliminates noise from low-probability branching

---

## 4. Empirical Results

| Metric            | Before      | After       | Improvement    |
| ----------------- | ----------- | ----------- | -------------- |
| System Throughput | 0.68695     | 0.74589     | +8.6%          |
| Total Runtime     | 379.94 s    | 349.91 s    | −30.03 s       |
| Unit Tests Passed | ✅ 1135/1135 | ✅ 1135/1135 | No regressions |
| Valgrind Check    | ✅           | ✅           | Clean          |

---

## 5. Validation Logs

- ✅ [`tests_before.txt`](https://github.com/user-attachments/files/21012976/tests_before.txt)
- ✅ [`tests_after.txt`](https://github.com/user-attachments/files/21012981/tests_after.txt)
- ✅ [`valgrind_before.txt`](https://github.com/user-attachments/files/21012978/valgrind_before.txt)
- ✅ [`valgrind_after.txt`](https://github.com/user-attachments/files/21012983/valgrind_after.txt)

---

## 6. Conclusion

This patch addresses over-branching in `TrinomialTree` construction under small time steps—critical in Hull-White-based lattice models. The entropy-based pruning is:

- Mathematically sound
- Lightweight
- Backward-safe